Refactor configuration settings to ensure environment variables are correctly parsed as integers. Updated MAX_FILE_SIZE and ACCESS_TOKEN_EXPIRE_MINUTES in settings and authentication routes for consistent type handling. Removed unused background tasks parameter in the upload endpoint for cleaner code.

+++ config/settings.py
@@ -11,7 +11,7 @@ UPLOAD_DIR = Path(os.getenv("UPLOAD_DIR", "uploads"))
 UPLOAD_DIR.mkdir(exist_ok=True)
 
 # File size limit (10MB)
-MAX_FILE_SIZE = os.getenv("MAX_FILE_SIZE", 10 * 1024 * 1024)
+MAX_FILE_SIZE = int(os.getenv("MAX_FILE_SIZE", 10 * 1024 * 1024))
 
 # Database Configuration
 DATABASE_URL = os.getenv(
@@ -22,7 +22,7 @@ DATABASE_URL = os.getenv(
 # JWT Configuration
 SECRET_KEY = os.getenv("SECRET_KEY")
 ALGORITHM = "HS256"
-ACCESS_TOKEN_EXPIRE_MINUTES = os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", 30)
+ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", 30))
 
 # LLM Configuration
 LLM_MODEL = os.getenv("LLM_MODEL", "gemini").lower()
